% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.outline.R
\name{na.outline}
\alias{na.outline}
\title{A Cat Function}
\usage{
na.outline(data, plot = FALSE)
}
\arguments{
\item{data}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
na.outline(Hitters,plot = TRUE)
}
\keyword{cats}
