% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R
\name{outline}
\alias{outline}
\title{A Cat Function}
\usage{
outline(data)
}
\arguments{
\item{data}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
outline(Hitters,plot = TRUE)
}
\keyword{cats}
