% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underlimit.R
\name{underlimit}
\alias{underlimit}
\title{Get observations under the limit}
\usage{
underlimit(data, limit)
}
\arguments{
\item{data}{Dataset, matrix or dataframe.}

\item{limit}{The lower limit of the desired missing rate (percentage).}
}
\description{
This function deletes observations above the missing percentage and returns new data.
}
\examples{
Hitters \%>\% underlimit(20)
}
\keyword{limit}
\keyword{missing,}
\keyword{under}
