% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withnomvoutlier.R
\name{withnomvoutlier}
\alias{withnomvoutlier}
\title{Get observations without multivariate outliers}
\usage{
withnomvoutlier(data, keepoutlier = FALSE)
}
\arguments{
\item{data}{Dataset, matrix or dataframe.}

\item{keepoutlier}{if TRUE, keeps outlier observations in a separate dataframe and returns a list including both.}
}
\description{
This function deletes multivariate outlier observations which are detected by robust principal component analysis and returns new data.
}
\examples{
Hitters \%>\% withnomvoutlier()
}
\keyword{outlier}
