% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.outline.R
\name{outlier.outline}
\alias{outlier.outline}
\title{Outlier Outline}
\usage{
outlier.outline(data)
}
\arguments{
\item{data}{Dataset, matrix or dataframe.}

\item{plot_show}{Logical argument for plotting outliers.}
}
\description{
This function allows you a report about outliers of your data.
}
\examples{
outlier.outline(Hitters,plot_show = TRUE)
}
\keyword{outlier}
